Extract token creation to separate file
