+++ src/mappings/colony.ts
@@ -1,14 +1,13 @@
-import { BigInt, log, Address } from '@graphprotocol/graph-ts'
+import { log, Address } from '@graphprotocol/graph-ts'
 
 import { IColony, DomainAdded, PaymentAdded, PaymentPayoutSet, ColonyMetadata, TokensMinted } from '../../generated/templates/Colony/IColony'
 
-import { Token as TokenContract } from '../../generated/templates/Token/Token'
-
-import { Token, Colony, Domain, Payment, FundingPotPayout, FundingPot } from '../../generated/schema'
+import { Colony, Domain, Payment, FundingPotPayout, FundingPot } from '../../generated/schema'
 
 import { OneTxPayment } from '../../generated/schema'
 
 import { handleEvent } from './event'
+import { createToken } from './token'
 
 export function handleDomainAdded(event: DomainAdded): void {
   let domain = new Domain(event.address.toHex() + '_domain_' +  event.params.domainId.toString())
@@ -53,25 +52,8 @@ export function handlePaymentPayoutSet(event: PaymentPayoutSet): void{
   }
   fundingPot.save()
 
-  let token = Token.load(fundingPotPayout.token)
-  if (token == null){
-    token = new Token(fundingPotPayout.token);
-    let t = TokenContract.bind(Address.fromString(fundingPotPayout.token))
+  createToken(fundingPotPayout.token)
 
-    let decimals = t.try_decimals()
-    if (decimals.reverted){
-      token.decimals = BigInt.fromI32(18)
-    } else {
-      token.decimals = BigInt.fromI32(decimals.value)
-    }
-    let symbol = t.try_symbol()
-    if (!symbol.reverted){
-      token.symbol = symbol.value
-    }
-
-    token.save()
-
-  }
   handleEvent("PaymentPayoutSet(address,uint256,address,uint256)", event, event.address)
 }
