+++ src/mappings/colonyNetwork.ts
@@ -8,9 +8,9 @@ import {
   ExtensionInstalled
 } from '../../generated/ColonyNetwork/IColonyNetwork'
 
-import { Colony, Domain, Token } from '../../generated/schema'
-import { Token as TokenContract } from '../../generated/templates/Token/Token'
+import { Colony, Domain } from '../../generated/schema'
 import { Colony as ColonyTemplate, OneTxPayment as OneTxPaymentTemplate } from '../../generated/templates'
+import { createToken } from './token'
 
 export function handleColonyAdded(event: ColonyAdded): void {
   let rootDomain = new Domain(event.params.colonyAddress.toHex() + '_domain_1')
@@ -27,24 +27,7 @@ export function handleColonyAdded(event: ColonyAdded): void {
   }
 
   let tokenAddress = event.params.token.toHexString()
-  let token = Token.load(tokenAddress)
-  if (token == null){
-    token = new Token(tokenAddress);
-    let t = TokenContract.bind(Address.fromString(tokenAddress))
-
-    let decimals = t.try_decimals()
-    if (decimals.reverted){
-      token.decimals = BigInt.fromI32(18)
-    } else {
-      token.decimals = BigInt.fromI32(decimals.value)
-    }
-    let symbol = t.try_symbol()
-    if (!symbol.reverted){
-      token.symbol = symbol.value
-    }
-
-    token.save()
-  }
+  createToken(tokenAddress)
 
   colony.colonyChainId = event.params.colonyId
   colony.token = tokenAddress
